AP_L1_Control: prefer into-wind when doing waypoint u-turn
New param NAVL1_ANGLE_THR, default 120, in degrees. This value is the angle threshold where the wind vector is considered to force a turn direction which favors into-the-wind. Into-wind is favored when turn angle for next waypoint is greater than this value.

In an example of a 175 degree u-turn, without wind we take the shortest path which is to the right. With a left to right cross-wind we would have normally taken the shorter run-path (right) but then get blown downwind which causes a larger crosstrack error than anticipated resulting in a longer ground-path and longer line acquisition. This also cases with-wind flight briefly dropping our airspeed. When taking a cross-wind into the direction decision we will prefer to turn into the wind, a left. In the case of a very strong wind we would essentially spin in place creating a much shorter ground path as well as maintaining a safer airspeed.